+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to FairRoot will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## x.x.0 - xxxx-xx-xx
+
+### Breaking Changes
+
+* No streamer is generated for `FairRootManager` any more
+
 ## 19.0.0 - 2024-05-17
 
 ### Breaking Changes